refactoring coming along
